[Feat] #10 - FireBase 주입 및 DataSource 연동 완료

+++ app/src/main/java/com/example/zupzup/data/datasource/StoreDataSource.kt
@@ -5,5 +5,5 @@ import com.example.zupzup.data.datamodel.Store
 interface StoreDataSource {
     suspend fun getStoreList(): Result<List<Store>>
 
-    suspend fun getStoreDetailById(storeId: Int): Result<Store>
+    suspend fun getStoreDetailById(storeId: Long): Result<Store>
 }
